Added Repository unit tests and modified Rating model class

+++ app/src/main/java/com/rajit/unittestingexample/model/Rating.kt
@@ -1,6 +1,6 @@
 package com.rajit.unittestingexample.model
 
 data class Rating(
-    val rate: Float,
+    val rate: Double,
     val count: Int
 )
